ci: make build-environment workflow callable
Make the build-environment callable. This will be used to create tagged
container images when the aquarist-labs/s3gw repo is tagged and a
project release is triggered.

The aim is that the container tagged with :latest contains the latest
changes in this repo. But there are also containers tagged with the
project's release version containing the correct contents for that
specific release.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>