MADR 0001: Reorganize Documentation
Propose new organization for the documentation. The purpose of
reorganizing the documentation is to make it easier for people
unfamiliar with the project to find relevant information.
In order to do that a new hirarchy is proposed. The hirarchy
specifically has the needs of users who are not contributing, users who
are contributing feedback and then code contributors in that order in
mind.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>