container: implement stopsignal handling
Implement correct stopsignal handling for the container image.
The http-server process responds quickly and well defined to a SIGINT
with a shutdown. This is much better than letting the container runtime
clean up the process and all resources after timing out when terminating
the container with SIGTERM (the default).
This allows for quicker shutdowns aiding any HA setups with pacemaker
and generally being much nicer to work with for operators.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>