tests: create regression test suite of s3tests
- Split out known-good s3tests for use as a regression test suite.

Since not all s3tests pass at the moment, the suite as a whole will
always fail. This makes it unsuitable for use in a CI pipeline. To
combat this, the known-good tests are split out into their own test
suite to be used as regression tests in CI pipelines. The main use case
for this is as merge criteria for PRs.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>