build: base on openSUSE Leap 15.4 instead of Tumbleweed
This relies on the filesystems:ceph:s3gw project on OBS to provide
dependencies that are not present (or are too old) in Leap 15.4:

  https://build.opensuse.org/project/show/filesystems:ceph:s3gw

At the time of writing, most of the packages present there (boost,
fmt, gperfools, pprof and thrift) are linked to the current versions
from openSUSE:Factory and I've tied them to those versions, so if
(for example) Factory later updates to a newer version of boost, we
will still remain on version 1.80.0, i.e. there should be no
surprise dependecy updates.  The exception to this is libsqliteorm,
which links to home:jluis/libsqliteorm.

I've also switched to slightly older versions of libicu and luarocks
because the older versions of those two present in Leap don't seem
to cause any trouble for the build, and it saves us rebuilding newer
versions.

Finally, we need to explicitly set CMAKE_C_COMPILER=gcc-11 and
CMAKE_CXX_COMPILER=g++-11, because while gcc 11 *is* available on
Leap, the default on Leap is version 7.

Fixes: https://github.com/aquarist-labs/s3gw/issues/162
Signed-off-by: Tim Serong <tserong@suse.com>