tools: change build-radosgw.sh to use ceph's repo script
This commit goes in conjunction with https://github.com/aquarist-labs/ceph/pull/68

In order to make ceph repo independant from s3gw-tools and be able to
run the unittests and other tests the build script has been moved to the
ceph repository.

The new script also offer the possibility to set a user defined build
script (via the new env variable BUILD_SCRIPT) in case we have special needs for
particular tests or to extend the ci on the ceph repository.

If, for example, we need to add a new flag when building the gateway, or
a new test it will be done in the ceph repository and the change will be
tested in its own ci in the same pull request.

This way the build container remains the same being just a holder of the
needed environment, but the how we build the gateways is ruled by the ceph
repository.

This also moves the tests to the ceph repository, so we can also run
them in any ceph's PR without depending on s3gw-tools.

It also creates 2 new containers for running the `radosgw` and its unit
tests in the `ceph` repo CI.

Those containers are mere base for running those processes and the
binaries and libraries are mounted. That way we don't need to fully
create a container with the binaries and we can reuse them.

Fixes: https://github.com/aquarist-labs/s3gw/issues/127

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>