ci: s3tests runner log from container
Make the s3tests runner script collect logs, when running the s3gw
within the runtime container. Previously, logs of the s3gw would only be
collected when running in the host environment. However this isn't how
the tests are run in the CI system, so in order to be able to collect
logs there, the container runtime has to be instructed to write out its
logs to a file, from where the logs can be collected.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>