ci: s3tests print summary and fix docker logs
- Fix docker log collection.

Docker and Podman use different options to
configure logging. These options aren't compatible and there aren't
always exactly equivalent options present. Therefore the s3tests
runner script needs to implement two different ways of collecting logs
from inside a container.

- Add FORCE_DOCKER option

To test the ability to run with Docker on a system that has (the
preferred) Podman available, there must be an environment variable
option FORCE_DOCKER={ON,OFF} to force the use of Docker despite the
presence of Podman

- Add DEBUG option

To improve debugging capabilities, running the script in an environment
where $DEBUG is not empty causes tracing to be enabled. This avoids
having to call `bash -x s3tests-runner.sh`

- Print results summary

At the end of running tests, not only should the results be collected in
a file for later analysis, but also printed to console. This makes it
much easier to know what succeeded/failed both for local tests as well
as test workflows. Both the count and a list of tests which
passed/failed is printed to stdout.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>