ci: s3tests runner, migrate to tox
- Migrate to `tox`

Upstream ceph/s3-tests is using `tox` instead of `nosetests` to run
tests. Fix the resulting compatibility issues with the s3tests runner
script.

- Fix failure report logic

The logic to detect failed tests was flawed, as it would return `1`
(there are failed tests) even, when there weren't any failed tests. This
is because the usage og `grep` in that context is wrong. Rewriting that
logic with bash-internals fixes this issue.

- Fix log collection for Podman

Podman log collection needs to rename the temporary log file _after_ the
container has been removed for sure, since it otherwise might not be
closed and thus result in errors.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>