ci: unify container build
Unify the container build by creating a multi-stage contaienr which
takes a copy of the ceph subrepo and builds the radosgw binary with the
sfs backend inside it. The next stage then creates a runtime container
capable of executing the resulting artefacts.
The total result of this is that building the s3gw container is no more
compliacted that a `docker build -t s3gw .`

Requires-pr: aquarist-labs/ceph#130
Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>