ci: simplify release workflow
Simplify release workflow by utilizing the unified Dockerfile.
Artefact generation now is just one build job instead of three. The
tests are uniformly attached after the build process for the container
has completed.
The process now mimics a more sensible

  build -> test -> draft-release

pattern instead of the previous pattern with tests intermingled with
different build stages.
This also removes the requirement that the aquarist-labs/ceph repo must
contain certain tags corresponding to the tags that trigger the release
workflow in this repo.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>