ci: add unit tests to release workflow
Add unit testing step to release workflow. This utilizes the docker
cache to avoid building all image layers twice. First it builds the
shared layers for the build environment, then the unit tests. These unit
tests are then run and if they succeed, based on the cached layers the
release container is built. After that integration tests continue as was
previously the case.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>