The Traefik ingress is misconfigured, Access-Control-Expose-Headers must be set to '*'
The `accessControlExposeHeaders` property is set to `*` now to expose all headers by default. To fix the mentioned issue adding the `X-Amz-Object-Lock-Mode` and `X-Amz-Object-Lock-Retain-Until-Date` would be sufficient, but the problem might appear again in the future, but with a different header that is processed by the AWS SDK but is not exposed by the ingress.

References:
- https://github.com/aquarist-labs/s3gw-charts/pull/89
- https://github.com/aquarist-labs/s3gw/issues/410

Signed-off-by: Volker Theile <vtheile@suse.com>