ci: run s3tests in parallel
Speed up s3tests by running them in parallel.
Each test is running against its own instance of the s3gw anyways, since
this is needed to limit the blast radius of crashes. Thus mutiple tests
can be executed concurrently as they do not depend on each other and do
not interfere with each other on different instances of the s3gw.
GNU parallel provides the job queue and flock is used to synchronize
result collection.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>