s3tr: Replace new-success/new-failures with summary and excuses.csv
Remove new-successes/new-failures commands.
Add new summary command that either prints stats or with a excuses.csv
compares the test results to an excuse file. This inverts the previous
logic where we had a file with test that must pass. We now have a list
with tests that are ok to fail plus URLs/excuses

Signed-off-by: Marcel Lauhoff <marcel.lauhoff@suse.com>