Dockerfile: retry `zypper install` up to three times
We've variously had problems with `zypper install` timing out both
in CI, and sometimes on local developer systems.  When this happens
it kills the entire build, which is extremely irritating.  It seems
we're not the first project to hit this problem (see
https://github.com/openSUSE/zypper/issues/420).

My hope here is that by retrying `zypper install` if the first call
fails due to a dodgy mirror (or whatever), the second or third call
will succeed.  At best this amounts to a no-op.  At worst, it'll
completely fail more slowly than it did before.

Signed-off-by: Tim Serong <tserong@suse.com>