Dockerfile: use bci 15.5 as base image
Note that in order for the build to work we have to add the
leap 15.5 update repos to the buildenv container, so that
the various necessary devel packages are available.

This commit also explicitly installs lua53-devel rather than
the generic lua-devel, which would otherwise pull in lua54-devel
which later causes problems at runtime (15.5 includes both
lua53 and lua54, but only the former is installed by default).

Fixes: https://github.com/aquarist-labs/s3gw/issues/771
Signed-off-by: Tim Serong <tserong@suse.com>