Dockerfile: Turn on ENABLE_GIT_VERSION
ENABLE_GIT_VERSION works fine when ceph is checked out normally,
but not when it's a submodule of s3gw that's copied into the Docker
build environment (".git" is a file that refers to a parent directory
that doesn't exist in this case).  Happily, we can cheat by mounting
the s3gw directory into the buildenv, then we can run a couple of git
commands inside *there* to extract the commit hash and version tags
ceph needs then save them to /srv/ceph/src/.git_version where they'll
be picked up by the ceph build.

Fixes: https://github.com/aquarist-labs/s3gw/issues/431
Signed-off-by: Tim Serong <tserong@suse.com>